requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
black==21.12b0
cfgv==3.3.1
charset-normalizer==2.0.10
click==8.0.3
deepgram-sdk==0.2.1
distlib==0.3.4
ffmpeg-python==0.2.0
filelock==3.4.2
frozenlist==1.3.0
future==0.18.2
identify==2.4.4
idna==3.3
multidict==5.2.0
mypy-extensions==0.4.3
nodeenv==1.6.0
numpy==1.22.1
opencv-python==4.5.5.62
pathspec==0.9.0
platformdirs==2.4.1
pre-commit==2.17.0
python-dotenv==0.19.2
PyYAML==6.0
six==1.16.0
toml==0.10.2
tomli==1.2.3
typing_extensions==4.0.1
virtualenv==20.13.0
websockets==10.1
yarl==1.7.2
